requirements.txt
pandas==0.24.2
numpy==1.16.4
flask==1.1.1
nltk==3.4.4
gunicorn==20.0.4
gevent==1.4.0
requests==2.5.3
scikit-learn==0.21.2
elasticsearch==7.0.4
Pillow
flask-cors==3.0.8